Create Birthday_Reminder.py
Birthday Reminder
The Birthday Reminder is a simple Python application that allows users to store and retrieve birthdays of their friends. This application provides a user-friendly interface to add and view birthdays conveniently.

Features:
Show Birthday: Users can look up birthdays by entering the name of their friend. If the friend's birthday is stored in the app, it will display the corresponding date; otherwise, it will prompt the user to add the birthday to the list.

Add to Birthday List: Users can add new friends and their birthdays to the reminder list.

Exit: Users can choose to exit the application when they are done using it.

Usage:
Run the script.
Choose an option from the menu:
1: Show Birthday
2: Add to Birthday List
3: Exit
How to Contribute:
If you want to contribute to this project, feel free to fork the repository and submit a pull request with your changes. Any enhancements, bug fixes, or additional features are welcome!